refactor(evo2-sae): move steering clamp hook into the evo2 recipe
Relocates the delta-clamp steering hook out of the shared sae library and
into the recipe, so this PR's code is entirely under recipes/evo2/:
  sae/src/sae/steering.py      -> recipes/evo2/src/evo2_sae/steering.py
  sae/tests/test_steering.py   -> recipes/evo2/tests/test_clamp.py
core.py + the moved test import from evo2_sae.steering. The recipe still
consumes the shared sae lib for TopKSAE + sae.eval (unchanged).
